Обработка исключений и Spring REST.

+++ src/main/java/ru/job4j/chat/repository/MessageRepository.java
@@ -6,12 +6,12 @@ import ru.job4j.chat.model.Message;
 
 import java.util.Optional;
 
-public interface MessageRepository extends CrudRepository<Message, Long> {
+public interface MessageRepository extends CrudRepository<Message, Integer> {
     @Override
     @EntityGraph(attributePaths = "person")
     Iterable<Message> findAll();
 
     @Override
     @EntityGraph(attributePaths = "person")
-    Optional<Message> findById(Long id);
+    Optional<Message> findById(Integer id);
 }
