+++ src/main/java/ru/job4j/chat/service/RoomServiceImpl.java
@@ -25,7 +25,7 @@ public class RoomServiceImpl implements RoomService {
     }
 
     @Override
-    public Optional<Room> findById(long id) {
+    public Optional<Room> findById(int id) {
         return roomRep.findById(id);
     }
